Add files via upload
Bootloader for the Fairlight CMI lightpen/USB mouse, built using Microchip MPLAB v8.63
The interface uses a PIC32 microcontroller. The PIC32 Flash memory is divided into 2 sections:
1. Bootloader (this code)
2. Application (lightpen emulator)
At startup this code checks if a USB mass storage device is attached, and if it contains a firmware update file.
If it does, the file is copied into Flash as the Application. The Bootloader then jumps to the Application code.
If if does not, the Bootloader just jumps to the Application code.