Add files via upload
Application for the Fairlight CMI lightpen/USB mouse, built using Microchip MPLAB v8.63
The interface uses a PIC32 microcontroller. The PIC32 Flash memory is divided into 2 sections:
1. Bootloader
2. Application (this code)
This code tracks the position of a connected USB mouse and converts that position into signals which emulate signals generated by a Fairlight CMI lightpen.
These lightpens are rare and difficult to repair. With this interface, a modern USB mouse can be used to drive the Fairlight UI.