Add files via upload
Version 2 of the Fairlight lightpen / USB mouse interface.
This version mounts inside the CMI, in an unused mainframe slot.
+5V power is taken from the ribbon cable to the front panel.
For a Series I or II, the "Series_I_power" board is needed, because the ribbon cable from the front panel has a different number of pins.
2 lasercut "wings" are used to hold the board in place in the mainframe slot.
A rear plate mounts to the bottom pan of the CMI, over 4 of the 8 I/O openings. You can lasercut this plate or use the .fpd file to order a cool aluminum one from Front Panel Express.